log diagnostics to stderr
